updated solution
