+++ Greed_Game_spec.rb
@@ -117,3 +117,4 @@ end
        
     end 
  end
+
